backend-mock/requirements.txt
Django>=4.2
djangorestframework>=3.14
visine-common
pytest-django>=4.5
